+++ gnarrator/config/keys.json
@@ -1,5 +1,6 @@
 {
     "FULL_SCREEN": "\u00ba", 
     "REGION": "f21",
-    "CLEAR": "esc"
+    "CLEAR": "esc",
+    "SMALL_N_QUICK": "-"
 }
